requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
greenlet==2.0.1
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
urllib3==1.26.13
